Move logger to logger directory
